requirements.txt
pybind11[global]
pyyaml
scipy
termcolor
imviz
matplotlib
